Stop the dispatcher from rendering failed matches
If a match fails, the stack is empty, no actions to call.

The check for the empty stack was at the end of the dispatch logic, after
the call to the renderer, so even with failed matches, the render logic
was still being called.

This change also makes sure that the stash is clean after a failed match.

Signed-off-by: Pedro Melo <melo@simplicidade.org>