make sure that all byte order marks are matched
The keys of %$BOM must be matched in correct order because the UTF-16LE BOM
will also match the UTF-32LE BOM, so UTF-32LE must be tried first.
Currently it worked correctly on my version of perl but the order of
unsorted hash keys are not guaranteed to be the same all the time
and for sorted keys (UTF-16LE before UTF-32LE) json.t fails.