Remove mysterious comments and describe the real problem in Changes
The earlier change from "use lib" to "BEGIN { unshift @INC, … }" was
meant to solve https://github.com/kraih/mojo/issues/462. Alas, "use lib"
cleverly adds more directories to @INC than requested: here it added the
core lib directory before site_perl, making newer versions of dual-life
modules invisible.

The purpose of the FindBin and "use lib" was to allow the script to find
the modules that it was distributed with (script/my_app → lib/MyApp.pm),
but the comment made it seem that the FindBin+BEGIN were needed to make
dual-life modules work. Without comment, the code is unremarkable and
easy to understand.