Assume content length is len buffer when no content-length header
Not that this is probably correct, but it works for me when abusing
Mojo::Message to my own ends.

I'm not sure if there is a way to get a message through to be parsed
without the header if so, this seemed the safest guess and better
than not parsing anything.