Mojo::Pointer doesn't throw an error when JSON Pointer is missing a starting '/' (against RFC 6901)
Hi!

I have a JSON object, which had a bug:

`
{
  "type": "object",
  "properties": {
    "listContentNumber": { "$ref": "../definitions.json#listContentNumber" },
    "biblionumber": { "$ref": "../definitions.json#/biblionumber" },
    "itemnumber": { "$ref": "../definitions.json#/itemnumber" },
    "borrowernumber": { "$ref" : "../definitions.json#/borrowernumber" }
  }
}
`

can you spot it?
Mojo::Pointer couldn't, and I hope you can accept my solution to improve it.

Looking into the ietf specification for JSON Pointer:
    http://tools.ietf.org/html/rfc6901#section-3
if the pointer has a 'reference-token', it MUST be prepended by a '/'-character.

  **>   json-pointer = *( "/" reference-token )   <**
  * any number of "/" + reference-token -tokens are allowed, but if the reference-token is defined,
  * the prepending "/" is mandatory and the JSON Pointer implementation SHOULD throw an error.

for example this is invalid:
    { "$ref": "../definitions.json#listContentNumber" }
whereas this is valid:
    { "$ref": "../definitions.json#/listContentNumber" }

If the reference-token doesn't 'start' with '/' then it returns the whole data-object!
It should check for pointer validity and die instead.