Make Mojo::Template support Devel::Cover
Add support for Devel::Cover in Mojo::Template, by monkey patching the compiled
template into its own namespace. This way the previously-anonymous sub exists
in the symbol table, and therefore Devel::Cover can see it.

As suggested by @haarg in pjcj/Devel--Cover#31

A path attribute is also added, to provide the path to the template instead of
using the name, which doesn't identify the template file.