fix "split to @_ is deprecated" warning
On Perl 5.10, perl is incorrectly interpreting any split not assigned to
an array as a split intended to replace @_. Since this behavior was
deprecated, Perl is warning about it. This warning is causing tests in
downstream modules (like my Mercury application) which check for
warnings to fail.

I can't find when this deprecation was added (before 5.6), but the
feature was removed from Perl in 5.12, and the warning fixed to only
apply to split called in void context.

The only syntax that prevents this warning is assigning the result of
split to an array. Disabling the category of this warning ("syntax", as
reported in perldiag) does not seem to prevent this warning, but
disabling all warnings does.