fix uninitialized warnings in stream
After `undef $self->{buffer}`, we can no longer call `length` on it in
Perl 5.10 (later Perls more usefully treat `undef` as empty without
warning). But, according to #1256, we can't just set `$self->{buffer}`
to the empty string without causing Perl to hold on to all the memory it
allocated for our buffer. So, we have to first `undef`, then set to
a defined value to prevent warnings from `length` on Perl 5.10.