Update default SSL/TLS key to support TLS 1.2+
This updates the default key used for local HTTPS server testing to
support TLS 1.2+. An RSA key of 4096 bits is created, with SHA256
digest support, with duration and Common Name as before.

As of Debian 10 "buster" the default system supported TLS level is 1.2,
and as a result, the current server.crt certificate using RSA:1024 with SHA1
digest support is not supported and will no longer work "out of the box".

With Apple, Google, Microsoft, and Mozilla announcing plans to deprecate
TLS 1.0 and 1.1 in early 2020, and Linux distributions updating their
baseline TLS support, it seems prudent to update this key so that it
continues to work for developers.