Handle :scope cases where we need more than current subtree
In cases like ':scope + p', we cannot do the search strictly in the
current subtree, as the p resides outside of it.

To circumvent this problem, "absolutize" the selector (as the RFC
calls it) if the selector is relative, and search from the root.

This is only done when necessary (i.e. any selector mentions ":root" or
":scope"), as it is slower to traverse the whole tree, rather than only
the current subtree.

Addresses #1405