Mojo::Util#secure_compare fix leak of string length #1599
By immediately returning when the two strings are not the
same length, the function allows an attacker to guess the
length of the secret string using timing attacks.

This change uses a fix by @jberger, based on [1].

This also updates the documentation to emphasize that the
secret string should be the second argument (although it's
not critical).

[1] https://github.com/vadimdemedes/secure-compare/blob/master/index.js