remove workaround
Previously, the default log level of `Mojo::Log` was `debug`. This is now
either the value of `MOJO_LOG_LEVEL` or `debug`. But since `Mojo::Test`
sets `MOJO_LOG_LEVEL` to the value `fatal` in the side effect, the test
would fail without the adjustment in line 29.

For me it looks in such a way that one wanted to create here simply in
the test the condition to be able to compare by Handle the log outputs.
And that there so far `debug` came out as level, although it is executed
in a test and thus should be `fatal`, is somehow unexpected.

For this reason the level for this test should be predefined.