doc: Cleanup ipv4 address expansion from short form.
Commit e3af2e7b4b59d85367fe7c30c19b76498e105482 removed all short form
ipv4 network addresses to address compat issues. One such fix added too
many octets however, resulting in the invalid address of 172.16.0.0.0/12
(the previous was a short-form 172.16.0/12 that required adjustment).

Fix by removing the extra octet to leave 172.16.0.0/12.

I ran a scan of the entire Mojolicious repository looking for other
"five octets" addresses and there was nothing except a deliberate DNS
hostname meant 1.1.1.1.1.1 in the testsuite, presumably to validate that
it is not accidentally treated as an IP address.