Add a 'before_app_start' event.
This event is always called before the application starts, regardless
of whether it is invoked via morbo, is started in daemon mode from the
command line, via hypnotoad in pre-forking mode, as a CGI application
or as a PSGI.

In pre-forked children, it avoids having to say:

    Mojo::IOLoop->next_tick(sub ($ioloop) {
        app->log->info("Worker $$ star...ALL GLORY TO THE HYPNOTOAD!");
    });

to get this special behaviour.  It is way nicer to be able to say:

    app->hook(before_app_start => sub ($app) {
        $app->log->info("App in $$ star...ALL GLORY TO MOJOLICIOUS!");
    });

to get that behaviour.

It is more consistent with the over hooks, and also it is a better
abstraction than having to know the internals of the IOLoop to be
able to get such a hooking point in pre-forked processes.

Now the application has a unique and consistent entry point where it can
initialize its global state, regardless of how it is being launched.