doc: reword testing guide paragraph
The commit 2b5293a14290f1e0b679b527534d5d1396ec624b added the
Mojolicious::Lite testing snippet that loads the application from a
file as an alternative to the snippet before that. However, the
paragraph that explains the details for both snippets explicitly
mentions the 'Frogs' application, which is used only in the first
snippet, creating a possible confusion for some readers that may wonder
if that explanation is also valid for the Mojolicious::Lite snippet.

This patch removes the explicitly mention of 'Frogs' from the paragraph
and also make change the 'This object initializes' to 'The object
initializes' to give a more generic meaning, including both snippets.