fix error reporting when loading applications with compile errors
When switching from loading an application from using require to using
do, the error was still trying to be captured after a surrounding eval.
This error would always be empty, because the error from the do was not
rethrown. It would still result in an undef application, so an error
would be thrown, but it would not include the actual compile error.

Change the do call inside the eval to throw an error if no application
is returned. The error is then recaught outside the eval. Also add a
test that shows that the real compilation error is included in the
output.

Fixes #2110